+++ src/PortalBase/Builder/Builder/ExplorerBuilder.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Portal\Base\Builder\Builder;
 
+use Heptacom\HeptaConnect\Portal\Base\Builder\Support\BuilderPriorityTrait;
 use Heptacom\HeptaConnect\Portal\Base\Builder\Token\ExplorerToken;
 
 class ExplorerBuilder
 {
+    use BuilderPriorityTrait;
+
     public function __construct(
         private ExplorerToken $token
     ) {