+++ src/PortalBase/Builder/FlowComponent.php
@@ -40,6 +40,8 @@ final class FlowComponent implements LoggerAwareInterface
 {
     private LoggerInterface $logger;
 
+    private ?int $defaultPriority = null;
+
     /**
      * @var ExplorerToken[]
      */
@@ -176,8 +178,10 @@ final class FlowComponent implements LoggerAwareInterface
      */
     public function buildExplorers(): iterable
     {
-        foreach (self::$explorerTokens as $key => $explorerToken) {
-            yield new Explorer($explorerToken);
+        foreach (self::$explorerTokens as $key => $token) {
+            $priority = $token->getPriority() ?? $this->defaultPriority;
+
+            yield $priority => new Explorer($token);
             unset(self::$explorerTokens[$key]);
         }
     }
@@ -187,12 +191,14 @@ final class FlowComponent implements LoggerAwareInterface
      */
     public function buildEmitters(): iterable
     {
-        foreach (self::$emitterTokens as $key => $emitterToken) {
-            if ($emitterToken->getRun() instanceof \Closure && $emitterToken->getBatch() instanceof \Closure) {
+        foreach (self::$emitterTokens as $key => $token) {
+            if ($token->getRun() instanceof \Closure && $token->getBatch() instanceof \Closure) {
                 $this->logImplementationConflict('Emitter', 'run', 'batch');
             }
 
-            yield new Emitter($emitterToken);
+            $priority = $token->getPriority() ?? $this->defaultPriority;
+
+            yield $priority => new Emitter($token);
             unset(self::$emitterTokens[$key]);
         }
     }
@@ -202,12 +208,14 @@ final class FlowComponent implements LoggerAwareInterface
      */
     public function buildReceivers(): iterable
     {
-        foreach (self::$receiverTokens as $key => $receiverToken) {
-            if ($receiverToken->getRun() instanceof \Closure && $receiverToken->getBatch() instanceof \Closure) {
+        foreach (self::$receiverTokens as $key => $token) {
+            if ($token->getRun() instanceof \Closure && $token->getBatch() instanceof \Closure) {
                 $this->logImplementationConflict('Receiver', 'run', 'batch');
             }
 
-            yield new Receiver($receiverToken);
+            $priority = $token->getPriority() ?? $this->defaultPriority;
+
+            yield $priority => new Receiver($token);
             unset(self::$receiverTokens[$key]);
         }
     }
@@ -217,8 +225,10 @@ final class FlowComponent implements LoggerAwareInterface
      */
     public function buildStatusReporters(): iterable
     {
-        foreach (self::$statusReporterTokens as $key => $statusReporterToken) {
-            yield new StatusReporter($statusReporterToken);
+        foreach (self::$statusReporterTokens as $key => $token) {
+            $priority = $token->getPriority() ?? $this->defaultPriority;
+
+            yield $priority => new StatusReporter($token);
             unset(self::$statusReporterTokens[$key]);
         }
     }
@@ -255,13 +265,16 @@ final class FlowComponent implements LoggerAwareInterface
                 }
             }
 
-            yield new HttpHandler($httpHandlerToken);
+            $priority = $httpHandlerToken->getPriority() ?? $this->defaultPriority;
+
+            yield $priority => new HttpHandler($httpHandlerToken);
             unset(self::$httpHandlerTokens[$key]);
         }
     }
 
     public function reset(): void
     {
+        $this->defaultPriority = null;
         self::$explorerTokens = [];
         self::$emitterTokens = [];
         self::$receiverTokens = [];
@@ -272,6 +285,11 @@ final class FlowComponent implements LoggerAwareInterface
     public function setLogger(LoggerInterface $logger): void
     {
         $this->logger = $logger;
+    }
+
+    public function setDefaultPriority(int $defaultPriority): void
+    {
+        $this->defaultPriority = $defaultPriority;
     }
 
     private function logImplementationConflict(string $builder, string $method, string $dropped): void