+++ src/PortalBase/Builder/Support/BuilderPriorityTrait.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Portal\Base\Builder\Support;
+
+trait BuilderPriorityTrait
+{
+    public function priority(int $priority): self
+    {
+        $this->token->setPriority($priority);
+
+        return $this;
+    }
+}