+++ src/PortalBase/Builder/Token/HttpHandlerToken.php
@@ -4,8 +4,12 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Portal\Base\Builder\Token;
 
+use Heptacom\HeptaConnect\Portal\Base\Builder\Support\TokenPriorityTrait;
+
 class HttpHandlerToken
 {
+    use TokenPriorityTrait;
+
     private ?\Closure $run = null;
 
     private ?\Closure $options = null;