+++ src/PortalBase/Web/Http/HttpHandlerStackIdentifier.php
@@ -25,6 +25,7 @@ final class HttpHandlerStackIdentifier implements FlowComponentStackIdentifierIn
         return $this->path;
     }
 
+    #[\ReturnTypeWillChange]
     public function jsonSerialize()
     {
         return [