+++ test/PortalBase/Builder/FlowComponentTest.php
@@ -98,6 +98,8 @@ final class FlowComponentTest extends TestCase
             ConfigurationContract::class => $config,
         ][$id] ?? null);
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -105,6 +107,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->options(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -112,6 +116,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->get(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -119,6 +125,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->post(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -126,6 +134,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->put(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -133,6 +143,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->patch(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -140,6 +152,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->delete(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -166,6 +180,8 @@ final class FlowComponentTest extends TestCase
         $logger->expects(static::exactly(6))->method('warning');
         $flowComponent->setLogger($logger);
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn () => $response)->options(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -173,6 +189,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn () => $response)->get(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -180,6 +198,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn () => $response)->post(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -187,6 +207,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn () => $response)->put(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -194,6 +216,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn () => $response)->patch(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -201,6 +225,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn () => $response)->delete(static fn () => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -224,6 +250,8 @@ final class FlowComponentTest extends TestCase
             ConfigurationContract::class => $config,
         ][$id] ?? null);
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn () => '');
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -236,6 +264,8 @@ final class FlowComponentTest extends TestCase
             static::assertSame('Short-noted HttpHandler failed in run to return Psr\Http\Message\ResponseInterface', $throwable->getMessage());
         }
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->options(static fn () => '');
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -248,6 +278,8 @@ final class FlowComponentTest extends TestCase
             static::assertSame('Short-noted HttpHandler failed in options to return Psr\Http\Message\ResponseInterface', $throwable->getMessage());
         }
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->get(static fn () => '');
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -260,6 +292,8 @@ final class FlowComponentTest extends TestCase
             static::assertSame('Short-noted HttpHandler failed in get to return Psr\Http\Message\ResponseInterface', $throwable->getMessage());
         }
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->post(static fn () => '');
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -272,6 +306,8 @@ final class FlowComponentTest extends TestCase
             static::assertSame('Short-noted HttpHandler failed in post to return Psr\Http\Message\ResponseInterface', $throwable->getMessage());
         }
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->put(static fn () => '');
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -284,6 +320,8 @@ final class FlowComponentTest extends TestCase
             static::assertSame('Short-noted HttpHandler failed in put to return Psr\Http\Message\ResponseInterface', $throwable->getMessage());
         }
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->patch(static fn () => '');
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -296,6 +334,8 @@ final class FlowComponentTest extends TestCase
             static::assertSame('Short-noted HttpHandler failed in patch to return Psr\Http\Message\ResponseInterface', $throwable->getMessage());
         }
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->delete(static fn () => '');
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -324,6 +364,8 @@ final class FlowComponentTest extends TestCase
             ConfigurationContract::class => $config,
         ][$id] ?? null);
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn (ConfigurationContract $config) => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -347,6 +389,8 @@ final class FlowComponentTest extends TestCase
             ConfigurationContract::class => $config,
         ][$id] ?? null);
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn (ServerRequestInterface $request) => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -354,6 +398,8 @@ final class FlowComponentTest extends TestCase
         static::assertCount(1, $handlers);
         static::assertSame($response, $handlers[0]->handle($request, $response, $context, $stack));
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn (RequestInterface $request) => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -377,6 +423,8 @@ final class FlowComponentTest extends TestCase
             ConfigurationContract::class => $config,
         ][$id] ?? null);
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn (ResponseInterface $r) => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());
@@ -400,6 +448,8 @@ final class FlowComponentTest extends TestCase
             ConfigurationContract::class => $config,
         ][$id] ?? null);
 
+        $flowComponent->reset();
+        $flowComponent->setDefaultPriority(0);
         $handlerBuilder = FlowComponent::httpHandler('foobar');
         $handlerBuilder->run(static fn (HttpHandleContextInterface $c) => $response);
         $handlers = \iterable_to_array($flowComponent->buildHttpHandlers());