+++ src/StorageBase/CHANGELOG.md
@@ -63,6 +63,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.9.7.0] - 2024-02-10
+
+### Fixed
+
+- Fix deprecation notice `Creation of dynamic property` by explicitly declaring property `\Heptacom\HeptaConnect\Storage\Base\Bridge\Support\AbstractSingletonStorageFacade::$portalNodeStorageSetAction`
+
 ## [0.9.6.0] - 2023-07-10
 
 ## [0.9.5.0] - 2023-05-27